Vagrant: Enable Vagrant in WSL
Vagrant needs to be running the same verson on both WSL and on Windows,
and the user needs to be in the Hyper-V Administrators group in order to
create/destroy and manage VMs in Hyper-V.

Because the vagrant package must be at a specific version, also set the
other packages in Windows to be at state: latest instead of state:
present. This will prevent packages (like Vagrant) from being updated.